+++ api/khairat-api-go/model/person.go
@@ -1,7 +1,7 @@
 package model
 
 type Person struct {
-	Id       int    `gorm:"column:id;primaryKey" json:"id"`
+	Id       int64  `gorm:"column:id;primaryKey" json:"id"`
 	Name     string `gorm:"type:varchar(128)" json:"name"`
 	IcNumber string `gorm:"type:varchar(12)" json:"icNumber"`
 	Address  string `gorm:"type:varchar(256)" json:"address"`
