+++ api/khairat-api-go/model/tag.go
@@ -1,7 +1,7 @@
 package model
 
 type Tag struct {
-	Id   int    `gorm:"column:id;primaryKey" json:"id"`
+	Id   int64  `gorm:"column:id;primaryKey" json:"id"`
 	Name string `gorm:"type:varchar(12)" json:"name"`
 }
